requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
PyJWT==1.6.4
pytest==4.0.0
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
